商品詳細画面カテゴリ修正、サービスをインターフェイスと実装分け、管理者編集修正、プロフィール編集修正、アカウント新規登録修正実施
