+++ ProductInventorySystem/src/main/resources/templates/login.html
@@ -4,7 +4,7 @@
     <title>Login</title>
 </head>
 <body>
-	    <h1>管理者ログイン</h1>
+    <h1>管理者ログイン</h1>
     <div th:if="${param.error}">
         <p>Invalid username and password.</p>
     </div>
@@ -22,5 +22,6 @@
             <input type="password" name="password" required/>
         </div>
         <button type="submit">Login</button>
+    </form>
 </body>
 </html>
