+++ ProductInventorySystem/src/main/resources/templates/product-details.html
@@ -33,6 +33,14 @@
         <label>大カテゴリ名:</label>
         <span th:text="${largeCategoryName}"></span>
     </div>
+    <div>
+        <label>中カテゴリ名:</label>
+        <span th:text="${subcategoryName}"></span>
+    </div>
+    <div>
+        <label>小カテゴリ名:</label>
+        <span th:text="${smallCategoryName}"></span>
+    </div>
     <a th:href="@{/products}">戻る</a>
 </body>
 </html>
