+++ ProductInventorySystem/src/main/resources/templates/register-manager.html
@@ -20,7 +20,7 @@
         </div>
         <div>
             <label>電話番号:</label>
-            <input type="text" th:field="*{phoneNumber}" required />
+            <input type="text" th:field="*{phoneNumber}" pattern="\d{10,11}" title="ハイフンなしで入力してください" required />
         </div>
         <div>
             <label>パスワード:</label>
